+++ package.json
@@ -45,8 +45,6 @@
     "ethers": "^5.5.3",
     "express": "^4.17.1",
     "firebase": "^9.5.0",
-    "google-protobuf": "^3.19.3",
-    "grpc-web": "^1.3.0",
     "human-standard-token-abi": "^2.0.0",
     "keccak": "byterose/keccak#stream-update",
     "lodash.debounce": "^4.0.8",
