--- src/components/bridge/EvmWidget.vue
@@ -1,351 +1,0 @@
-<template>
-  <div>
-    <div class="container">
-      <div class="widget" :class="isDarkTheme && 'widget-dark'">
-        <div class="row">
-          <div class="currency">
-            <span class="label">{{ $t('from') }}</span>
-            <div class="chain" :class="isDarkTheme && 'chain-dark'" @click="openModal('src')">
-              <div>
-                <img v-if="srcChain" :src="srcChain.icon" alt="src-chain-logo" class="chain-logo" />
-              </div>
-              <span v-if="srcChain" class="chain-name">
-                {{ srcChain ? getChainName(srcChain.id) : '' }}
-              </span>
-              <span class="under-arrow">▼</span>
-            </div>
-          </div>
-          <div>
-            <div class="input-row" :class="isDarkTheme && 'input-row-dark'">
-              <input
-                :value="amount"
-                inputmode="decimal"
-                type="number"
-                min="0"
-                pattern="^[0-9]*(\.)?[0-9]*$"
-                placeholder="0"
-                @input="inputHandler"
-              />
-              <button
-                class="max-button"
-                :class="isDarkTheme && 'max-button-dark'"
-                @click="toMaxAmount"
-              >
-                {{ $t('bridge.max') }}
-              </button>
-              <div
-                class="token-selector"
-                :class="isDarkTheme && 'token-selector-dark'"
-                @click="openModal('token')"
-              >
-                <img
-                  v-if="selectedToken"
-                  :src="selectedToken.symbol === 'USDT' ? logoUsdt : selectedToken.icon"
-                  alt="token-logo"
-                  class="token-logo"
-                />
-                <span>{{ selectedToken ? selectedToken.symbol : '' }}</span>
-                <span>▼</span>
-              </div>
-            </div>
-            <div class="information label">
-              <!-- Todo: Un-comment after fix the debounce issue -->
-              <!-- <span class="label"> ${{ $n(usdValue) }}</span> -->
-              <div />
-              <div class="balance">
-                <p>{{ $t('bridge.balance') }}</p>
-                <p>
-                  {{
-                    selectedTokenBalance
-                      ? formatDecimals({
-                          amount: selectedTokenBalance,
-                          decimals: 6,
-                        })
-                      : 0
-                  }}
-                  {{ selectedToken ? selectedToken.symbol : '' }}
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="reverse" :class="isDarkTheme && 'reverse-dark'" @click="reverseChain">
-          <span class="reverse-button" :class="isDarkTheme && 'reverse-button-dark'">↑↓</span>
-        </div>
-        <div class="row">
-          <div class="currency">
-            <span class="label">{{ $t('to') }}</span>
-            <div class="chain" :class="isDarkTheme && 'chain-dark'" @click="openModal('dest')">
-              <div>
-                <img
-                  v-if="destChain"
-                  :src="destChain.icon"
-                  alt="destChain-chain-logo"
-                  class="chain-logo"
-                />
-              </div>
-              <span class="chain-name">{{ destChain ? getChainName(destChain.id) : '' }}</span>
-              <span class="under-arrow">▼</span>
-            </div>
-          </div>
-          <div>
-            <div class="estimation-row" :class="isDarkTheme && 'estimation-row-dark'">
-              <span class="label">{{ $t('estimated') }}</span>
-              <span class="estimated-value">{{
-                amount && quotation && quotation.estimated_receive_amt > 0
-                  ? quotation.estimated_receive_amt
-                  : ''
-              }}</span>
-            </div>
-          </div>
-        </div>
-
-        <EvmBridgeButtons
-          :bridge="bridge"
-          :handle-approve="handleApprove"
-          :open-select-modal="openSelectModal"
-          :is-disabled-bridge="isDisabledBridge"
-          :is-approval-needed="isApprovalNeeded"
-          :selected-network="selectedNetwork"
-          :src-chain-id="srcChain.id"
-        />
-      </div>
-      <div class="remarks-container" :class="isDarkTheme && 'remarks-container-dark'">
-        <div class="remarks-left">
-          <p>{{ $t('bridge.bridgeRate') }}</p>
-          <p>{{ $t('bridge.fee') }}</p>
-          <p>{{ $t('bridge.timeOfArrival') }}</p>
-          <p>{{ $t('bridge.minAmount') }}</p>
-          <p>{{ $t('bridge.maxAmount') }}</p>
-        </div>
-        <div class="remarks-right">
-          <div class="rate">
-            <p class="rate-mobile">{{ quotation ? quotation.bridge_rate : 1 }}</p>
-            <div class="rate-md">
-              <div class="rate-column">
-                <span>{{
-                  $t('bridge.rateSymbol', {
-                    symbol: selectedToken ? selectedToken.symbol : '',
-                    value: 1,
-                  })
-                }}</span>
-                <img
-                  v-if="srcChain"
-                  :src="srcChain.icon"
-                  alt="src-chain-logo"
-                  class="chain-logo-mini"
-                />
-              </div>
-              <span>=</span>
-              <div class="rate-column">
-                <span>{{
-                  $t('bridge.rateSymbol', {
-                    symbol: selectedToken ? selectedToken.symbol : '',
-                    value: quotation ? quotation.bridge_rate : 1,
-                  })
-                }}</span>
-                <img
-                  v-if="destChain"
-                  :src="destChain.icon"
-                  alt="destChain-chain-logo"
-                  class="chain-logo-mini"
-                />
-              </div>
-            </div>
-          </div>
-          <div class="quotation-value-currency">
-            <span class="quotation-value">{{
-              quotation
-                ? quotation.base_fee.includes('e') // eg: '1e-8' (WBTC)
-                  ? amount
-                    ? Number(amount) - Number(quotation.estimated_receive_amt)
-                    : 0
-                  : quotation.base_fee
-                : 0
-            }}</span>
-            <span class="quotation-currency">{{ selectedToken ? selectedToken.symbol : '' }}</span>
-          </div>
-          <p class="time">{{ $t('bridge.time', { from: '5', to: '20' }) }}</p>
-          <div class="quotation-value-currency">
-            <span class="quotation-value">{{ quotation ? quotation.minAmount : 0 }}</span>
-            <span class="quotation-currency">{{ selectedToken ? selectedToken.symbol : '' }}</span>
-          </div>
-          <div class="quotation-value-currency">
-            <span class="quotation-value">{{
-              $n(quotation ? Number(quotation.maxAmount) : 0)
-            }}</span>
-            <span class="quotation-currency">{{ selectedToken ? selectedToken.symbol : '' }}</span>
-          </div>
-        </div>
-      </div>
-      <div class="provider">
-        <a
-          href="https://cbridge.celer.network/#/transfer"
-          target="_blank"
-          rel="noopener noreferrer"
-          class="powered-link"
-        >
-          Powered by Celer Network
-        </a>
-      </div>
-
-      <ModalToken
-        v-if="modal === 'token'"
-        v-model:isOpen="modal"
-        :close-modal="closeModal"
-        :select-token="selectToken"
-        :tokens="tokens"
-        :selected-token="selectedToken"
-        :src-chain-id="srcChain.id"
-        :modal="modal"
-      />
-
-      <ModalChain
-        v-if="modal === 'src'"
-        v-model:isOpen="modal"
-        :close-modal="closeModal"
-        :select-chain="selectChain"
-        :chains="srcChains"
-        :modal="modal"
-        :selected-chain="srcChain"
-      />
-
-      <ModalChain
-        v-if="modal === 'dest'"
-        v-model:isOpen="modal"
-        :close-modal="closeModal"
-        :select-chain="selectChain"
-        :chains="destChains"
-        :modal="modal"
-        :selected-chain="destChain"
-      />
-
-      <modal-connect-wallet
-        v-if="walletModalName === WalletModalOption.SelectWallet"
-        :set-wallet-modal="setWalletModal"
-        :set-close-modal="closeWalletModal"
-        :is-evm-only="true"
-      />
-
-      <ModalInstallWallet
-        v-if="walletModalName === WalletModalOption.NoExtension"
-        :set-close-modal="closeWalletModal"
-        :selected-wallet="selectedWallet"
-      />
-    </div>
-  </div>
-</template>
-
-<script lang="ts">
-import { useMeta } from 'quasar';
-import { getChainName, formatDecimals } from 'src/c-bridge';
-import { useCbridge, useConnectWallet } from 'src/hooks';
-import { calUsdAmount } from 'src/hooks/helper/price';
-import { useStore } from 'src/store';
-import { computed, defineComponent } from 'vue';
-import ModalConnectWallet from '../balance/modals/ModalConnectWallet.vue';
-import ModalInstallWallet from '../balance/modals/ModalInstallWallet.vue';
-import ModalChain from './modals/ModalChain.vue';
-import ModalToken from './modals/ModalToken.vue';
-import EvmBridgeButtons from './EvmBridgeButtons.vue';
-
-export default defineComponent({
-  components: {
-    ModalChain,
-    ModalToken,
-    ModalConnectWallet,
-    ModalInstallWallet,
-    EvmBridgeButtons,
-  },
-  setup() {
-    useMeta({ title: 'EVM Bridge' });
-
-    const store = useStore();
-    const isDarkTheme = computed(() => store.getters['general/theme'] === 'DARK');
-
-    const {
-      srcChain,
-      destChain,
-      srcChains,
-      destChains,
-      chains,
-      tokens,
-      modal,
-      amount,
-      selectedToken,
-      quotation,
-      isApprovalNeeded,
-      selectedTokenBalance,
-      selectedNetwork,
-      isDisabledBridge,
-      usdValue,
-      closeModal,
-      openModal,
-      selectChain,
-      selectToken,
-      reverseChain,
-      inputHandler,
-      toMaxAmount,
-      handleApprove,
-      bridge,
-    } = useCbridge();
-
-    const {
-      WalletModalOption,
-      modalConnectWallet,
-      modalName: walletModalName,
-      selectedWallet,
-      isH160,
-      setCloseModal: closeWalletModal,
-      setWalletModal,
-      openSelectModal,
-    } = useConnectWallet();
-
-    const logoUsdt = 'https://s2.coinmarketcap.com/static/img/coins/64x64/825.png';
-
-    return {
-      getChainName,
-      formatDecimals,
-      isDarkTheme,
-      srcChain,
-      destChain,
-      chains,
-      tokens,
-      modal,
-      amount,
-      destChains,
-      srcChains,
-      selectedToken,
-      quotation,
-      selectedTokenBalance,
-      logoUsdt,
-      selectedNetwork,
-      isDisabledBridge,
-      usdValue,
-      closeModal,
-      openModal,
-      selectChain,
-      selectToken,
-      reverseChain,
-      toMaxAmount,
-      bridge,
-      WalletModalOption,
-      modalConnectWallet,
-      walletModalName,
-      selectedWallet,
-      isH160,
-      closeWalletModal,
-      setWalletModal,
-      openSelectModal,
-      inputHandler,
-      handleApprove,
-      isApprovalNeeded,
-      calUsdAmount,
-    };
-  },
-});
-</script>
-
-<style lang="scss" scoped>
-@import './styles/evm-bridge.scss';
-</style>
