+++ src/components/bridge/evm/modals/ModalToken.vue
@@ -1,14 +1,17 @@
 <template>
-  <Modal :title="title" @click="closeModal">
+  <Modal title="Select a token" class="animate__animated animate__fadeIn" @click="closeModal">
     <template #content>
       <div class="tw-max-w-sm">
-        <Chain
-          v-for="(chain, index) in chains"
-          :key="index"
-          :chain="chain"
-          :select-chain="selectChain"
-          :selected-chain="selectedChain"
-        />
+        <div>
+          <Token
+            v-for="(token, index) in tokens"
+            :key="index"
+            :token-obj="token"
+            :selected-token="selectedToken"
+            :src-chain-id="srcChainId"
+            :select-token="selectToken"
+          />
+        </div>
       </div>
     </template>
   </Modal>
@@ -16,11 +19,11 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import Chain from './Chain.vue';
+import Token from './Token.vue';
 import Modal from 'src/components/common/Modal.vue';
 export default defineComponent({
   components: {
-    Chain,
+    Token,
     Modal,
   },
   props: {
@@ -28,20 +31,24 @@ export default defineComponent({
       type: Function,
       required: true,
     },
-    selectChain: {
+    selectToken: {
       type: Function,
       required: true,
     },
-    chains: {
+    tokens: {
       type: Array,
+      required: true,
+    },
+    selectedToken: {
+      type: Object,
+      required: true,
+    },
+    srcChainId: {
+      type: Number,
       required: true,
     },
     modal: {
       type: String,
-      required: true,
-    },
-    selectedChain: {
-      type: Object,
       required: true,
     },
   },
@@ -52,7 +59,3 @@ export default defineComponent({
   },
 });
 </script>
-
-<style lang="scss" scoped>
-@import '../styles/bridge-modal';
-</style>
