+++ src/i18n/en-US/index.ts
@@ -218,11 +218,11 @@ export default {
     balance: 'Available Balance:',
     approve: 'Approve',
     addWallet: 'Add to wallet',
-    bridgeRate: 'Bridge Rate',
+    bridgeRate: 'Bridge rate',
     fee: 'Fee',
-    timeOfArrival: 'Estimated Time of Arrival',
-    minAmount: 'Minimum bridge amount',
-    maxAmount: 'Maximum bridge amount',
+    timeOfArrival: 'Estimated time of arrival',
+    minAmount: 'Min. amount greater than',
+    maxAmount: 'Max. amount less than',
     time: '{from} ~ {to} mins',
     rateSymbol: '{value} {symbol} on ',
   },
